+++ example/lib/app/usecase_impl/CleanCoreExampleUseCaseImplExporter.dart
@@ -1,4 +1,0 @@
-library clean_core_example_usecase_impl_exporter;
-
-//export the usecases impl
-export 'ParentUseCaseImpl.dart';