+++ example/lib/repo/clean_core_examaple_repo_exporter.dart
@@ -1,7 +1,0 @@
-library clean_core_examaple_repo_exporter;
-
-export 'entities/Entities.dart';
-export 'module/CleanCoreExampleRepoModule.dart';
-export 'repo_external/CleanCoreExampleRepoExternalExporter.dart';
-export 'repo_external_impl/CleanCoreExampleRepoExternalImplExporter.dart';
-export 'repo_impl/RepoImplExporter.dart';