+++ example/lib/repo/clean_core_example_repo_exporter.dart
@@ -1,0 +1,7 @@
+library clean_core_examaple_repo_exporter;
+
+export 'entities/Entities.dart';
+export 'module/clean_core_example_repo_module.dart';
+export 'repo_external/clean_core_example_repo_internal_exporter.dart';
+export 'repo_external_impl/clean_core_example_repo_internal_impl_exporter.dart';
+export 'repo_impl/repo_impl_exporter.dart';