+++ example/lib/repo/repo_external/CleanCoreExampleRepoExternalExporter.dart
@@ -1,4 +1,0 @@
-library clean_core_example_repo_internal_exporter;
-
-//export the External repos
-export 'ParentRepoExternal.dart';