+++ example/lib/repo/repo_external_impl/CleanCoreExampleRepoExternalImplExporter.dart
@@ -1,4 +1,0 @@
-library clean_core_example_repo_internal_impl_exporter;
-
-//expor the External repos
-export 'ParentRepoExternalImpl.dart';