+++ example/lib/repo/repo_external_impl/clean_core_example_repo_internal_impl_exporter.dart
@@ -1,0 +1,4 @@
+library clean_core_example_repo_internal_impl_exporter;
+
+//expor the External repos
+export 'parent_repo_external_impl.dart';