+++ example/lib/repo/repo_impl/repo_impl_exporter.dart
@@ -1,0 +1,4 @@
+library repo_impl_exporter;
+
+///export the repos impl
+export 'parent_repo_impl.dart';