+++ lib/src/app/domain/CleanCoreDomainExporter.dart
@@ -1,7 +1,0 @@
-library clean_core_domain_exporter;
-
-//export the domains
-export 'DomainObject.dart';
-export 'BasicDomainObject.dart';
-
-export 'IntIdentifier.dart';