+++ lib/src/app/domain/clean_core_domain_exporter.dart
@@ -1,0 +1,7 @@
+library clean_core_domain_exporter;
+
+///export the domains
+export 'basic_domain_object.dart';
+export 'cloneable.dart';
+export 'domain_object.dart';
+export 'int_identifier.dart';